requirements.txt
gTTS==2.3.1
matplotlib==3.6.3
numpy==1.24.1
openai==0.26.5
pandas==1.5.2
Pillow==9.4.0
requests==2.28.2
streamlit==1.17.0
streamlit_authenticator==0.1.5
streamlit_chat==0.0.2.1
streamlit_extras==0.2.6
streamlit_lottie==0.0.3
defusedxml==0.7.1
extra-streamlit-components==0.1.56
matplotlib==3.6.3
matplotlib-inline==0.1.6
streamlit-aggrid==0.3.3
streamlit-chat==0.0.2.1